feat(storybook): 9개 컴포넌트 스토리 추가 + 셀프리뷰 이슈 반영
- stories/ 디렉토리로 위치 이동 (src/ui/components/ 에 두면 npm publish 시 포함됨)
- Storybook 기본값 변경: dark/purple → light/LG Lilac
- Input placeholder example@lge.com → user@example.com (OSS 대비)
- Text fixture 실명 → 역사적 인명으로 교체
- Toggle SettingsList '텔레메트리 수집' → '활동 데이터 수집'
- Button.tsx 내부 Spinner → ButtonSpinner (public export 이름 충돌 방지)
- .gitignore에 storybook-static/ 추가
