+++ .gitignore
@@ -3,3 +3,4 @@ node_modules/
 .omc/
 .host/
 package-lock.json
+storybook-static/
