--- src/ui/components/Spinner.stories.tsx
@@ -1,34 +1,0 @@
-import type { Meta, StoryObj } from "@storybook/react";
-import { Spinner } from "./Spinner.js";
-
-const meta: Meta<typeof Spinner> = {
-  title: "UI/Spinner",
-  component: Spinner,
-  argTypes: {
-    size: { control: "select", options: ["sm", "md", "lg"] },
-  },
-};
-export default meta;
-type Story = StoryObj<typeof Spinner>;
-
-export const Small: Story = {
-  args: { size: "sm" },
-};
-
-export const Medium: Story = {
-  args: { size: "md" },
-};
-
-export const Large: Story = {
-  args: { size: "lg" },
-};
-
-export const AllSizes: Story = {
-  render: () => (
-    <div style={{ display: "flex", gap: "1.5rem", alignItems: "center" }}>
-      <Spinner size="sm" />
-      <Spinner size="md" />
-      <Spinner size="lg" />
-    </div>
-  ),
-};
