+++ stories/Select.stories.tsx
@@ -1,4 +1,3 @@
-import React from "react";
 import type { Meta, StoryObj } from "@storybook/react";
 import { Select } from "../src/ui/components/Select.js";
 import { Text } from "../src/ui/components/Text.js";
@@ -6,36 +5,62 @@ import { Text } from "../src/ui/components/Text.js";
 const meta: Meta<typeof Select> = {
   title: "UI/Select",
   component: Select,
+  argTypes: {
+    disabled: { control: "boolean" },
+  },
 };
 export default meta;
+type Story = StoryObj<typeof Select>;
 
-const OPTIONS = (
-  <>
-    <option value="">선택하세요</option>
-    <option value="1">옵션 1</option>
-    <option value="2">옵션 2</option>
-    <option value="3">옵션 3</option>
-  </>
-);
+export const Default: Story = {
+  args: {
+    children: (
+      <>
+        <option value="">선택하세요</option>
+        <option value="gpt4o">GPT-4o</option>
+        <option value="claude">Claude</option>
+        <option value="gemini">Gemini</option>
+      </>
+    ),
+  },
+};
 
-export const All: StoryObj = {
+export const WithDefault: Story = {
+  args: {
+    defaultValue: "claude",
+    children: (
+      <>
+        <option value="gpt4o">GPT-4o</option>
+        <option value="claude">Claude</option>
+        <option value="gemini">Gemini</option>
+      </>
+    ),
+  },
+};
+
+export const Disabled: Story = {
+  args: {
+    disabled: true,
+    defaultValue: "claude",
+    children: (
+      <>
+        <option value="gpt4o">GPT-4o</option>
+        <option value="claude">Claude</option>
+        <option value="gemini">Gemini</option>
+      </>
+    ),
+  },
+};
+
+export const WithLabel: Story = {
   render: () => (
-    <div style={{ display: "flex", flexDirection: "column", gap: "1rem", maxWidth: "320px" }}>
-      <div>
-        <Text variant="label" style={{ marginBottom: "0.375rem" }}>기본</Text>
-        <Select>{OPTIONS}</Select>
-      </div>
-      <div>
-        <Text variant="label" style={{ marginBottom: "0.375rem" }}>비활성화</Text>
-        <Select disabled>{OPTIONS}</Select>
-      </div>
+    <div style={{ display: "flex", flexDirection: "column", gap: "0.25rem", width: "16rem" }}>
+      <Text variant="label" as="span">언어 모델</Text>
+      <Select defaultValue="claude">
+        <option value="gpt4o">GPT-4o</option>
+        <option value="claude">Claude</option>
+        <option value="gemini">Gemini</option>
+      </Select>
     </div>
   ),
-};
-
-export const Default: StoryObj<typeof Select> = {
-  render: () => <Select style={{ maxWidth: "320px" }}>{OPTIONS}</Select>,
-};
-export const Disabled: StoryObj<typeof Select> = {
-  render: () => <Select disabled style={{ maxWidth: "320px" }}>{OPTIONS}</Select>,
 };
