+++ stories/Spinner.stories.tsx
@@ -1,14 +1,29 @@
-import React from "react";
 import type { Meta, StoryObj } from "@storybook/react";
 import { Spinner } from "../src/ui/components/Spinner.js";
 
 const meta: Meta<typeof Spinner> = {
   title: "UI/Spinner",
   component: Spinner,
+  argTypes: {
+    size: { control: "select", options: ["sm", "md", "lg"] },
+  },
 };
 export default meta;
+type Story = StoryObj<typeof Spinner>;
 
-export const All: StoryObj = {
+export const Small: Story = {
+  args: { size: "sm" },
+};
+
+export const Medium: Story = {
+  args: { size: "md" },
+};
+
+export const Large: Story = {
+  args: { size: "lg" },
+};
+
+export const AllSizes: Story = {
   render: () => (
     <div style={{ display: "flex", gap: "1.5rem", alignItems: "center" }}>
       <Spinner size="sm" />
@@ -17,7 +32,3 @@ export const All: StoryObj = {
     </div>
   ),
 };
-
-export const Small: StoryObj<typeof Spinner> = { args: { size: "sm" } };
-export const Medium: StoryObj<typeof Spinner> = { args: { size: "md" } };
-export const Large: StoryObj<typeof Spinner> = { args: { size: "lg" } };
