+++ stories/Toggle.stories.tsx
@@ -1,29 +1,44 @@
-import React from "react";
 import type { Meta, StoryObj } from "@storybook/react";
 import { Toggle } from "../src/ui/components/Toggle.js";
-import { Text } from "../src/ui/components/Text.js";
 
 const meta: Meta<typeof Toggle> = {
   title: "UI/Toggle",
   component: Toggle,
+  argTypes: {
+    label: { control: "text" },
+    disabled: { control: "boolean" },
+  },
 };
 export default meta;
+type Story = StoryObj<typeof Toggle>;
 
-export const All: StoryObj = {
+export const Off: Story = {
+  args: { label: "자동 업데이트" },
+};
+
+export const On: Story = {
+  args: { label: "자동 업데이트", defaultChecked: true },
+};
+
+export const Disabled: Story = {
+  args: { label: "비활성 항목", disabled: true },
+};
+
+export const DisabledOn: Story = {
+  args: { label: "비활성 (활성화됨)", disabled: true, defaultChecked: true },
+};
+
+export const NoLabel: Story = {
+  args: { defaultChecked: true },
+};
+
+export const SettingsList: Story = {
   render: () => (
-    <div style={{ display: "flex", flexDirection: "column", gap: "0.875rem" }}>
-      <Toggle label="알림 허용" defaultChecked={false} />
-      <Toggle label="켜진 상태" defaultChecked={true} />
-      <Toggle label="비활성화 (꺼짐)" disabled defaultChecked={false} />
-      <Toggle label="비활성화 (켜짐)" disabled defaultChecked={true} />
-      <div>
-        <Text variant="label" style={{ marginBottom: "0.5rem" }}>레이블 없음</Text>
-        <Toggle defaultChecked={true} />
-      </div>
+    <div style={{ display: "flex", flexDirection: "column", gap: "0.875rem", width: "14rem" }}>
+      <Toggle label="자동 업데이트" defaultChecked />
+      <Toggle label="활동 데이터 수집" />
+      <Toggle label="크래시 리포트" defaultChecked />
+      <Toggle label="베타 기능" disabled />
     </div>
   ),
 };
-
-export const Off: StoryObj<typeof Toggle> = { args: { label: "알림 허용", defaultChecked: false } };
-export const On: StoryObj<typeof Toggle> = { args: { label: "알림 허용", defaultChecked: true } };
-export const Disabled: StoryObj<typeof Toggle> = { args: { label: "비활성화", disabled: true, defaultChecked: false } };
